requirements.txt
python-docx
Flask
azure-storage-blob
tensorflow==2.8.0
pandas
numpy

